Add a whisper feature for members on standby.
For this to work the EPGP officer should be in raid and the
auto-whisper feature should be enabled. In this mode the addon
interprets incoming whispers. If 'epgp standby' is received the sender
is added to the award list.  If 'epgp standby <name>' is received then
<name> is added to the award list.  After each award to the raid the
extra people are awarded the same EP. They are also removed from the
list and they are whispered the EP they received and also the fact
that they have been removed from the list. At this point, from here to
the next award they need to readd themselves to it.

This fixes issue 217.